requirements.txt
httpx-extensions>=0.1.1
orjson>=3.6
pydantic>=1.9.0
